requirements/base.txt
celery==5.2.7
Django==4.0.6
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.20.0
Pillow==9.2.0
psycopg2-binary==2.9.3
pytest==7.1.2
pytest-django==4.5.2
python-dotenv==0.20.0
redis==4.3.4
